Rename the beatmap language change handler.
